first credit register

+++ app/src/main/java/com/example/calculapp/data/sqlite/model/CreditModel.kt
@@ -2,7 +2,7 @@ package com.example.calculapp.data.sqlite.model
 
 data class CreditModel (
     val id: Int,
-    val userIdentificationNumber: Int,
+    val userIdentificationNumber: Long,
     val amountRequested: Int,
     val daysRequested: Int,
     val creditDate: String,
